resolve: make dashboard to a component and let AI create basic frameworks for different pages
